Reestructurando el proyecto, agregando parsers y comenzando creación de arbol B+

+++ src/main/java/com/mio/models/table/PropertyType.java
@@ -2,6 +2,5 @@ package com.mio.models.table;
 
 public enum PropertyType {
 
-    BYTE, SHORT, INT, DOUBLE, FLOAT, LONG, CHAR, BOOLEAN, STRING
-
+    BYTE, SHORT, INT, FLOAT, DOUBLE, LONG, CHAR, BOOLEAN, STRING
 }
